chatbot-service/requirements.txt
openai
flask
python-dotenv
requests